requirements.txt
Flask==3.0.0
pymongo==4.6.0
pandas>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
spacy>=3.7.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
requests>=2.31.0
